physioportal/requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-cors-headers==3.10.1
django-extensions==3.1.3
djangorestframework==3.12.4
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
